Move attribute mutations to attribute graphql module
