+++ saleor/graphql/attribute/descriptions.py
