+++ saleor/graphql/attribute/mutations.py
@@ -1,0 +1,517 @@
+import graphene
+from django.core.exceptions import ObjectDoesNotExist, ValidationError
+from django.db import transaction
+from django.utils.text import slugify
+
+from ...attribute import models as models
+from ...attribute.error_codes import AttributeErrorCode
+from ...core.exceptions import PermissionDenied
+from ...core.permissions import (
+    PageTypePermissions,
+    ProductPermissions,
+    ProductTypePermissions,
+)
+from ..core.inputs import ReorderInput
+from ..core.mutations import BaseMutation, ModelDeleteMutation, ModelMutation
+from ..core.types.common import AttributeError
+from ..core.utils import (
+    from_global_id_strict_type,
+    validate_slug_and_generate_if_needed,
+)
+from ..core.utils.reordering import perform_reordering
+from ..meta.deprecated.mutations import ClearMetaBaseMutation, UpdateMetaBaseMutation
+from ..product.types import Attribute, AttributeValue
+from .descriptions import AttributeDescriptions, AttributeValueDescriptions
+from .enums import AttributeInputTypeEnum, AttributeTypeEnum
+
+
+class AttributeValueCreateInput(graphene.InputObjectType):
+    name = graphene.String(required=True, description=AttributeValueDescriptions.NAME)
+
+
+class AttributeCreateInput(graphene.InputObjectType):
+    input_type = AttributeInputTypeEnum(description=AttributeDescriptions.INPUT_TYPE)
+    name = graphene.String(required=True, description=AttributeDescriptions.NAME)
+    slug = graphene.String(required=False, description=AttributeDescriptions.SLUG)
+    type = AttributeTypeEnum(description=AttributeDescriptions.TYPE, required=True)
+    values = graphene.List(
+        AttributeValueCreateInput, description=AttributeDescriptions.VALUES
+    )
+    value_required = graphene.Boolean(description=AttributeDescriptions.VALUE_REQUIRED)
+    is_variant_only = graphene.Boolean(
+        required=False, description=AttributeDescriptions.IS_VARIANT_ONLY
+    )
+    visible_in_storefront = graphene.Boolean(
+        description=AttributeDescriptions.VISIBLE_IN_STOREFRONT
+    )
+    filterable_in_storefront = graphene.Boolean(
+        description=AttributeDescriptions.FILTERABLE_IN_STOREFRONT
+    )
+    filterable_in_dashboard = graphene.Boolean(
+        description=AttributeDescriptions.FILTERABLE_IN_DASHBOARD
+    )
+    storefront_search_position = graphene.Int(
+        required=False, description=AttributeDescriptions.STOREFRONT_SEARCH_POSITION
+    )
+    available_in_grid = graphene.Boolean(
+        required=False, description=AttributeDescriptions.AVAILABLE_IN_GRID
+    )
+
+
+class AttributeUpdateInput(graphene.InputObjectType):
+    name = graphene.String(description=AttributeDescriptions.NAME)
+    slug = graphene.String(description=AttributeDescriptions.SLUG)
+    remove_values = graphene.List(
+        graphene.ID,
+        name="removeValues",
+        description="IDs of values to be removed from this attribute.",
+    )
+    add_values = graphene.List(
+        AttributeValueCreateInput,
+        name="addValues",
+        description="New values to be created for this attribute.",
+    )
+    value_required = graphene.Boolean(description=AttributeDescriptions.VALUE_REQUIRED)
+    is_variant_only = graphene.Boolean(
+        required=False, description=AttributeDescriptions.IS_VARIANT_ONLY
+    )
+    visible_in_storefront = graphene.Boolean(
+        description=AttributeDescriptions.VISIBLE_IN_STOREFRONT
+    )
+    filterable_in_storefront = graphene.Boolean(
+        description=AttributeDescriptions.FILTERABLE_IN_STOREFRONT
+    )
+    filterable_in_dashboard = graphene.Boolean(
+        description=AttributeDescriptions.FILTERABLE_IN_DASHBOARD
+    )
+    storefront_search_position = graphene.Int(
+        required=False, description=AttributeDescriptions.STOREFRONT_SEARCH_POSITION
+    )
+    available_in_grid = graphene.Boolean(
+        required=False, description=AttributeDescriptions.AVAILABLE_IN_GRID
+    )
+
+
+class AttributeMixin:
+    @classmethod
+    def check_values_are_unique(cls, values_input, attribute):
+        # Check values uniqueness in case of creating new attribute.
+        existing_values = attribute.values.values_list("slug", flat=True)
+        for value_data in values_input:
+            slug = slugify(value_data["name"], allow_unicode=True)
+            if slug in existing_values:
+                msg = (
+                    "Value %s already exists within this attribute."
+                    % value_data["name"]
+                )
+                raise ValidationError(
+                    {
+                        cls.ATTRIBUTE_VALUES_FIELD: ValidationError(
+                            msg, code=AttributeErrorCode.ALREADY_EXISTS
+                        )
+                    }
+                )
+
+        new_slugs = [
+            slugify(value_data["name"], allow_unicode=True)
+            for value_data in values_input
+        ]
+        if len(set(new_slugs)) != len(new_slugs):
+            raise ValidationError(
+                {
+                    cls.ATTRIBUTE_VALUES_FIELD: ValidationError(
+                        "Provided values are not unique.",
+                        code=AttributeErrorCode.UNIQUE,
+                    )
+                }
+            )
+
+    @classmethod
+    def clean_values(cls, cleaned_input, attribute):
+        """Clean attribute values.
+
+        Transforms AttributeValueCreateInput into AttributeValue instances.
+        Slugs are created from given names and checked for uniqueness within
+        an attribute.
+        """
+        values_input = cleaned_input.get(cls.ATTRIBUTE_VALUES_FIELD)
+
+        if values_input is None:
+            return
+
+        for value_data in values_input:
+            value_data["slug"] = slugify(value_data["name"], allow_unicode=True)
+            attribute_value = models.AttributeValue(**value_data, attribute=attribute)
+            try:
+                attribute_value.full_clean()
+            except ValidationError as validation_errors:
+                for field, err in validation_errors.error_dict.items():
+                    if field == "attribute":
+                        continue
+                    raise ValidationError({cls.ATTRIBUTE_VALUES_FIELD: err})
+        cls.check_values_are_unique(values_input, attribute)
+
+    @classmethod
+    def clean_attribute(cls, instance, cleaned_input):
+        try:
+            cleaned_input = validate_slug_and_generate_if_needed(
+                instance, "name", cleaned_input
+            )
+        except ValidationError as error:
+            error.code = AttributeErrorCode.REQUIRED.value
+            raise ValidationError({"slug": error})
+
+        return cleaned_input
+
+    @classmethod
+    def _save_m2m(cls, info, attribute, cleaned_data):
+        super()._save_m2m(info, attribute, cleaned_data)
+        values = cleaned_data.get(cls.ATTRIBUTE_VALUES_FIELD) or []
+        for value in values:
+            attribute.values.create(**value)
+
+
+class AttributeCreate(AttributeMixin, ModelMutation):
+    # Needed by AttributeMixin,
+    # represents the input name for the passed list of values
+    ATTRIBUTE_VALUES_FIELD = "values"
+
+    attribute = graphene.Field(Attribute, description="The created attribute.")
+
+    class Arguments:
+        input = AttributeCreateInput(
+            required=True, description="Fields required to create an attribute."
+        )
+
+    class Meta:
+        model = models.Attribute
+        description = "Creates an attribute."
+        error_type_class = AttributeError
+        error_type_field = "attribute_errors"
+
+    @classmethod
+    def perform_mutation(cls, _root, info, **data):
+        input = data.get("input")
+        # check permissions based on attribute type
+        if input["type"] == AttributeTypeEnum.PRODUCT_TYPE.value:
+            permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
+        else:
+            permissions = (PageTypePermissions.MANAGE_PAGE_TYPES_AND_ATTRIBUTES,)
+        if not cls.check_permissions(info.context, permissions):
+            raise PermissionDenied()
+
+        instance = models.Attribute()
+
+        # Do cleaning and uniqueness checks
+        cleaned_input = cls.clean_input(info, instance, input)
+        cls.clean_attribute(instance, cleaned_input)
+        cls.clean_values(cleaned_input, instance)
+
+        # Construct the attribute
+        instance = cls.construct_instance(instance, cleaned_input)
+        cls.clean_instance(info, instance)
+
+        # Commit it
+        instance.save()
+        cls._save_m2m(info, instance, cleaned_input)
+
+        # Return the attribute that was created
+        return AttributeCreate(attribute=instance)
+
+
+class AttributeUpdate(AttributeMixin, ModelMutation):
+    # Needed by AttributeMixin,
+    # represents the input name for the passed list of values
+    ATTRIBUTE_VALUES_FIELD = "add_values"
+
+    attribute = graphene.Field(Attribute, description="The updated attribute.")
+
+    class Arguments:
+        id = graphene.ID(required=True, description="ID of an attribute to update.")
+        input = AttributeUpdateInput(
+            required=True, description="Fields required to update an attribute."
+        )
+
+    class Meta:
+        model = models.Attribute
+        description = "Updates attribute."
+        permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
+        error_type_class = AttributeError
+        error_type_field = "attribute_errors"
+
+    @classmethod
+    def clean_remove_values(cls, cleaned_input, instance):
+        """Check if the values to be removed are assigned to the given attribute."""
+        remove_values = cleaned_input.get("remove_values", [])
+        for value in remove_values:
+            if value.attribute != instance:
+                msg = "Value %s does not belong to this attribute." % value
+                raise ValidationError(
+                    {
+                        "remove_values": ValidationError(
+                            msg, code=AttributeErrorCode.INVALID
+                        )
+                    }
+                )
+        return remove_values
+
+    @classmethod
+    def _save_m2m(cls, info, instance, cleaned_data):
+        super()._save_m2m(info, instance, cleaned_data)
+        for attribute_value in cleaned_data.get("remove_values", []):
+            attribute_value.delete()
+
+    @classmethod
+    def perform_mutation(cls, _root, info, id, input):
+        instance = cls.get_node_or_error(info, id, only_type=Attribute)
+
+        # Do cleaning and uniqueness checks
+        cleaned_input = cls.clean_input(info, instance, input)
+        cls.clean_attribute(instance, cleaned_input)
+        cls.clean_values(cleaned_input, instance)
+        cls.clean_remove_values(cleaned_input, instance)
+
+        # Construct the attribute
+        instance = cls.construct_instance(instance, cleaned_input)
+        cls.clean_instance(info, instance)
+
+        # Commit it
+        instance.save()
+        cls._save_m2m(info, instance, cleaned_input)
+
+        # Return the attribute that was created
+        return AttributeUpdate(attribute=instance)
+
+
+class AttributeDelete(ModelDeleteMutation):
+    class Arguments:
+        id = graphene.ID(required=True, description="ID of an attribute to delete.")
+
+    class Meta:
+        model = models.Attribute
+        description = "Deletes an attribute."
+        permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
+        error_type_class = AttributeError
+        error_type_field = "attribute_errors"
+
+
+class AttributeUpdateMeta(UpdateMetaBaseMutation):
+    class Meta:
+        model = models.Attribute
+        description = "Update public metadata for attribute."
+        permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
+        public = True
+        error_type_class = AttributeError
+        error_type_field = "attribute_errors"
+
+
+class AttributeClearMeta(ClearMetaBaseMutation):
+    class Meta:
+        description = "Clears public metadata item for attribute."
+        model = models.Attribute
+        permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
+        public = True
+        error_type_class = AttributeError
+        error_type_field = "attribute_errors"
+
+
+class AttributeUpdatePrivateMeta(UpdateMetaBaseMutation):
+    class Meta:
+        description = "Update public metadata for attribute."
+        model = models.Attribute
+        permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
+        public = False
+        error_type_class = AttributeError
+        error_type_field = "attribute_errors"
+
+
+class AttributeClearPrivateMeta(ClearMetaBaseMutation):
+    class Meta:
+        description = "Clears public metadata item for attribute."
+        model = models.Attribute
+        permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
+        public = False
+        error_type_class = AttributeError
+        error_type_field = "attribute_errors"
+
+
+def validate_value_is_unique(attribute: models.Attribute, value: models.AttributeValue):
+    """Check if the attribute value is unique within the attribute it belongs to."""
+    duplicated_values = attribute.values.exclude(pk=value.pk).filter(slug=value.slug)
+    if duplicated_values.exists():
+        raise ValidationError(
+            {
+                "name": ValidationError(
+                    f"Value with slug {value.slug} already exists.",
+                    code=AttributeErrorCode.ALREADY_EXISTS.value,
+                )
+            }
+        )
+
+
+class AttributeValueCreate(ModelMutation):
+    attribute = graphene.Field(Attribute, description="The updated attribute.")
+
+    class Arguments:
+        attribute_id = graphene.ID(
+            required=True,
+            name="attribute",
+            description="Attribute to which value will be assigned.",
+        )
+        input = AttributeValueCreateInput(
+            required=True, description="Fields required to create an AttributeValue."
+        )
+
+    class Meta:
+        model = models.AttributeValue
+        description = "Creates a value for an attribute."
+        permissions = (ProductPermissions.MANAGE_PRODUCTS,)
+        error_type_class = AttributeError
+        error_type_field = "attribute_errors"
+
+    @classmethod
+    def clean_input(cls, info, instance, data):
+        cleaned_input = super().clean_input(info, instance, data)
+        cleaned_input["slug"] = slugify(cleaned_input["name"], allow_unicode=True)
+        return cleaned_input
+
+    @classmethod
+    def clean_instance(cls, info, instance):
+        validate_value_is_unique(instance.attribute, instance)
+        super().clean_instance(info, instance)
+
+    @classmethod
+    def perform_mutation(cls, _root, info, attribute_id, input):
+        attribute = cls.get_node_or_error(info, attribute_id, only_type=Attribute)
+        instance = models.AttributeValue(attribute=attribute)
+        cleaned_input = cls.clean_input(info, instance, input)
+        instance = cls.construct_instance(instance, cleaned_input)
+        cls.clean_instance(info, instance)
+
+        instance.save()
+        cls._save_m2m(info, instance, cleaned_input)
+        return AttributeValueCreate(attribute=attribute, attributeValue=instance)
+
+
+class AttributeValueUpdate(ModelMutation):
+    attribute = graphene.Field(Attribute, description="The updated attribute.")
+
+    class Arguments:
+        id = graphene.ID(
+            required=True, description="ID of an AttributeValue to update."
+        )
+        input = AttributeValueCreateInput(
+            required=True, description="Fields required to update an AttributeValue."
+        )
+
+    class Meta:
+        model = models.AttributeValue
+        description = "Updates value of an attribute."
+        permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
+        error_type_class = AttributeError
+        error_type_field = "attribute_errors"
+
+    @classmethod
+    def clean_input(cls, info, instance, data):
+        cleaned_input = super().clean_input(info, instance, data)
+        if "name" in cleaned_input:
+            cleaned_input["slug"] = slugify(cleaned_input["name"], allow_unicode=True)
+        return cleaned_input
+
+    @classmethod
+    def clean_instance(cls, info, instance):
+        validate_value_is_unique(instance.attribute, instance)
+        super().clean_instance(info, instance)
+
+    @classmethod
+    def success_response(cls, instance):
+        response = super().success_response(instance)
+        response.attribute = instance.attribute
+        return response
+
+
+class AttributeValueDelete(ModelDeleteMutation):
+    attribute = graphene.Field(Attribute, description="The updated attribute.")
+
+    class Arguments:
+        id = graphene.ID(required=True, description="ID of a value to delete.")
+
+    class Meta:
+        model = models.AttributeValue
+        description = "Deletes a value of an attribute."
+        permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
+        error_type_class = AttributeError
+        error_type_field = "attribute_errors"
+
+    @classmethod
+    def success_response(cls, instance):
+        response = super().success_response(instance)
+        response.attribute = instance.attribute
+        return response
+
+
+class AttributeReorderValues(BaseMutation):
+    attribute = graphene.Field(
+        Attribute, description="Attribute from which values are reordered."
+    )
+
+    class Meta:
+        description = "Reorder the values of an attribute."
+        permissions = (ProductTypePermissions.MANAGE_PRODUCT_TYPES_AND_ATTRIBUTES,)
+        error_type_class = AttributeError
+        error_type_field = "attribute_errors"
+
+    class Arguments:
+        attribute_id = graphene.Argument(
+            graphene.ID, required=True, description="ID of an attribute."
+        )
+        moves = graphene.List(
+            ReorderInput,
+            required=True,
+            description="The list of reordering operations for given attribute values.",
+        )
+
+    @classmethod
+    def perform_mutation(cls, _root, info, attribute_id, moves):
+        pk = from_global_id_strict_type(
+            attribute_id, only_type=Attribute, field="attribute_id"
+        )
+
+        try:
+            attribute = models.Attribute.objects.prefetch_related("values").get(pk=pk)
+        except ObjectDoesNotExist:
+            raise ValidationError(
+                {
+                    "attribute_id": ValidationError(
+                        f"Couldn't resolve to an attribute: {attribute_id}",
+                        code=AttributeErrorCode.NOT_FOUND,
+                    )
+                }
+            )
+
+        values_m2m = attribute.values
+        operations = {}
+
+        # Resolve the values
+        for move_info in moves:
+            value_pk = from_global_id_strict_type(
+                move_info.id, only_type=AttributeValue, field="moves"
+            )
+
+            try:
+                m2m_info = values_m2m.get(pk=int(value_pk))
+            except ObjectDoesNotExist:
+                raise ValidationError(
+                    {
+                        "moves": ValidationError(
+                            f"Couldn't resolve to an attribute value: {move_info.id}",
+                            code=AttributeErrorCode.NOT_FOUND,
+                        )
+                    }
+                )
+            operations[m2m_info.pk] = move_info.sort_order
+
+        with transaction.atomic():
+            perform_reordering(values_m2m, operations)
+        attribute.refresh_from_db(fields=["values"])
+        return AttributeReorderValues(attribute=attribute)
