+++ saleor/graphql/product/enums.py
@@ -1,22 +1,9 @@
 import graphene
 
-from ...attribute import AttributeInputType, AttributeType
-from ..core.enums import to_enum
-
-AttributeInputTypeEnum = to_enum(AttributeInputType)
-AttributeTypeEnum = to_enum(AttributeType)
-
 
 class ProductAttributeType(graphene.Enum):
     PRODUCT = "PRODUCT"
     VARIANT = "VARIANT"
-
-
-class AttributeValueType(graphene.Enum):
-    COLOR = "COLOR"
-    GRADIENT = "GRADIENT"
-    URL = "URL"
-    STRING = "STRING"
 
 
 class StockAvailability(graphene.Enum):
