+++ saleor/graphql/product/tests/test_attributes.py
@@ -3,13 +3,10 @@ from unittest import mock
 
 import graphene
 import pytest
-from django.core.exceptions import ValidationError
 from django.db.models import Q
-from django.utils.text import slugify
 from graphene.utils.str_converters import to_camel_case
 
 from ....attribute import AttributeInputType, AttributeType
-from ....attribute.error_codes import AttributeErrorCode
 from ....attribute.models import (
     Attribute,
     AttributeProduct,
@@ -25,28 +22,12 @@ from ....product.models import (
     ProductType,
     ProductVariant,
 )
+from ...attribute.enums import AttributeTypeEnum, AttributeValueType
 from ...core.utils import snake_to_camel_case
 from ...tests.utils import get_graphql_content
-from ..enums import AttributeTypeEnum, AttributeValueType, ProductAttributeType
+from ..enums import ProductAttributeType
 from ..filters import filter_attributes_by_product_types
-from ..mutations.attributes import validate_value_is_unique
 from ..types.attributes import resolve_attribute_value_type
-
-
-def test_validate_value_is_unique(color_attribute):
-    value = color_attribute.values.first()
-
-    # a new value but with existing slug should raise an error
-    with pytest.raises(ValidationError):
-        validate_value_is_unique(color_attribute, AttributeValue(slug=value.slug))
-
-    # a new value with a new slug should pass
-    validate_value_is_unique(
-        color_attribute, AttributeValue(slug="spanish-inquisition")
-    )
-
-    # value that already belongs to the attribute shouldn't be taken into account
-    validate_value_is_unique(color_attribute, value)
 
 
 def test_get_single_attribute_by_pk(user_api_client, color_attribute_without_values):
@@ -451,792 +432,6 @@ def test_attributes_in_collection_query(
     expected_flat_attributes_data = list(expected_qs.values_list("slug", flat=True))
 
     assert flat_attributes_data == expected_flat_attributes_data
-
-
-CREATE_ATTRIBUTES_QUERY = """
-    mutation createAttribute(
-        $name: String!, $slug: String, $type: AttributeTypeEnum!,
-        $values: [AttributeValueCreateInput]
-    ){
-        attributeCreate(input: {
-            name: $name, values: $values, slug: $slug, type: $type,
-        }) {
-            errors {
-                field
-                message
-            }
-            attributeErrors {
-                field
-                message
-                code
-            }
-            attribute {
-                name
-                slug
-                type
-                values {
-                    name
-                    slug
-                }
-                productTypes(first: 10) {
-                    edges {
-                        node {
-                            id
-                        }
-                    }
-                }
-            }
-        }
-    }
-"""
-
-
-def test_create_attribute_and_attribute_values(
-    staff_api_client, permission_manage_product_types_and_attributes
-):
-    query = CREATE_ATTRIBUTES_QUERY
-
-    attribute_name = "Example name"
-    name = "Value name"
-    variables = {
-        "name": attribute_name,
-        "values": [{"name": name}],
-        "type": AttributeTypeEnum.PRODUCT_TYPE.name,
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    assert not content["data"]["attributeCreate"]["attributeErrors"]
-    data = content["data"]["attributeCreate"]
-
-    # Check if the attribute was correctly created
-    assert data["attribute"]["name"] == attribute_name
-    assert data["attribute"]["slug"] == slugify(
-        attribute_name
-    ), "The default slug should be the slugified name"
-    assert (
-        data["attribute"]["productTypes"]["edges"] == []
-    ), "The attribute should not have been assigned to a product type"
-
-    # Check if the attribute values were correctly created
-    assert len(data["attribute"]["values"]) == 1
-    assert data["attribute"]["type"] == AttributeTypeEnum.PRODUCT_TYPE.name
-    assert data["attribute"]["values"][0]["name"] == name
-    assert data["attribute"]["values"][0]["slug"] == slugify(name)
-
-
-def test_create_page_attribute_and_attribute_values(
-    staff_api_client, permission_manage_page_types_and_attributes
-):
-    query = CREATE_ATTRIBUTES_QUERY
-
-    attribute_name = "Example name"
-    name = "Value name"
-    variables = {
-        "name": attribute_name,
-        "values": [{"name": name}],
-        "type": AttributeTypeEnum.PAGE_TYPE.name,
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_page_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    assert not content["data"]["attributeCreate"]["attributeErrors"]
-    data = content["data"]["attributeCreate"]
-
-    # Check if the attribute was correctly created
-    assert data["attribute"]["name"] == attribute_name
-    assert data["attribute"]["slug"] == slugify(
-        attribute_name
-    ), "The default slug should be the slugified name"
-    assert (
-        data["attribute"]["productTypes"]["edges"] == []
-    ), "The attribute should not have been assigned to a product type"
-
-    # Check if the attribute values were correctly created
-    assert len(data["attribute"]["values"]) == 1
-    assert data["attribute"]["type"] == AttributeTypeEnum.PAGE_TYPE.name
-    assert data["attribute"]["values"][0]["name"] == name
-    assert data["attribute"]["values"][0]["slug"] == slugify(name)
-
-
-@pytest.mark.parametrize(
-    "input_slug, expected_slug",
-    (
-        ("my-slug", "my-slug"),
-        (None, "my-name"),
-        ("", "my-name"),
-        ("わたし-わ-にっぽん-です", "わたし-わ-にっぽん-です"),
-    ),
-)
-def test_create_attribute_with_given_slug(
-    staff_api_client,
-    permission_manage_product_types_and_attributes,
-    input_slug,
-    expected_slug,
-):
-    staff_api_client.user.user_permissions.add(
-        permission_manage_product_types_and_attributes
-    )
-    query = """
-        mutation createAttribute(
-            $name: String!, $slug: String, $type: AttributeTypeEnum!) {
-        attributeCreate(input: {name: $name, slug: $slug, type: $type}) {
-            attributeErrors {
-                field
-                message
-                code
-            }
-            attribute {
-                slug
-            }
-        }
-    }
-    """
-
-    attribute_name = "My Name"
-    variables = {
-        "name": attribute_name,
-        "slug": input_slug,
-        "type": AttributeTypeEnum.PRODUCT_TYPE.name,
-    }
-    content = get_graphql_content(staff_api_client.post_graphql(query, variables))
-
-    assert not content["data"]["attributeCreate"]["attributeErrors"]
-    assert content["data"]["attributeCreate"]["attribute"]["slug"] == expected_slug
-
-
-def test_create_attribute_value_name_and_slug_with_unicode(
-    staff_api_client, permission_manage_product_types_and_attributes
-):
-    query = CREATE_ATTRIBUTES_QUERY
-    name = "わたし わ にっぽん です"
-    slug = "わたし-わ-にっぽん-で"
-    variables = {
-        "name": name,
-        "slug": slug,
-        "type": AttributeTypeEnum.PRODUCT_TYPE.name,
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["attributeCreate"]
-    assert not data["attributeErrors"]
-    assert data["attribute"]["name"] == name
-    assert data["attribute"]["slug"] == slug
-
-
-@pytest.mark.parametrize(
-    "name_1, name_2, error_msg, error_code",
-    (
-        (
-            "Red color",
-            "Red color",
-            "Provided values are not unique.",
-            ProductErrorCode.UNIQUE,
-        ),
-        (
-            "Red color",
-            "red color",
-            "Provided values are not unique.",
-            ProductErrorCode.UNIQUE,
-        ),
-    ),
-)
-def test_create_attribute_and_attribute_values_errors(
-    staff_api_client,
-    name_1,
-    name_2,
-    error_msg,
-    error_code,
-    permission_manage_product_types_and_attributes,
-    product_type,
-):
-    query = CREATE_ATTRIBUTES_QUERY
-    variables = {
-        "name": "Example name",
-        "type": AttributeTypeEnum.PRODUCT_TYPE.name,
-        "values": [{"name": name_1}, {"name": name_2}],
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    errors = content["data"]["attributeCreate"]["errors"]
-    assert errors
-    assert errors[0]["field"] == "values"
-    assert errors[0]["message"] == error_msg
-
-    product_errors = content["data"]["attributeCreate"]["attributeErrors"]
-    assert product_errors[0]["code"] == error_code.name
-
-
-UPDATE_ATTRIBUTE_MUTATION = """
-    mutation updateAttribute(
-        $id: ID!, $name: String!, $addValues: [AttributeValueCreateInput]!,
-        $removeValues: [ID]!) {
-    attributeUpdate(
-            id: $id,
-            input: {
-                name: $name, addValues: $addValues,
-                removeValues: $removeValues}) {
-        errors {
-            field
-            message
-        }
-        attributeErrors {
-            field
-            message
-            code
-        }
-        attribute {
-            name
-            slug
-            values {
-                name
-                slug
-            }
-            productTypes(first: 10) {
-                edges {
-                    node {
-                        id
-                    }
-                }
-            }
-        }
-    }
-}
-"""
-
-
-def test_update_attribute_name(
-    staff_api_client, color_attribute, permission_manage_product_types_and_attributes
-):
-    query = UPDATE_ATTRIBUTE_MUTATION
-    attribute = color_attribute
-    name = "Wings name"
-    slug = attribute.slug
-    node_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    variables = {"name": name, "id": node_id, "addValues": [], "removeValues": []}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    attribute.refresh_from_db()
-    data = content["data"]["attributeUpdate"]
-    assert data["attribute"]["name"] == name == attribute.name
-    assert data["attribute"]["slug"] == slug == attribute.slug
-    assert data["attribute"]["productTypes"]["edges"] == []
-
-
-def test_update_attribute_remove_and_add_values(
-    staff_api_client, color_attribute, permission_manage_product_types_and_attributes
-):
-    query = UPDATE_ATTRIBUTE_MUTATION
-    attribute = color_attribute
-    name = "Wings name"
-    attribute_value_name = "Red Color"
-    node_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    attribute_value_id = attribute.values.first().id
-    value_id = graphene.Node.to_global_id("AttributeValue", attribute_value_id)
-    variables = {
-        "name": name,
-        "id": node_id,
-        "addValues": [{"name": attribute_value_name}],
-        "removeValues": [value_id],
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    attribute.refresh_from_db()
-    data = content["data"]["attributeUpdate"]
-    assert not data["attributeErrors"]
-    assert data["attribute"]["name"] == name == attribute.name
-    assert not attribute.values.filter(pk=attribute_value_id).exists()
-    assert attribute.values.filter(name=attribute_value_name).exists()
-
-
-def test_update_empty_attribute_and_add_values(
-    staff_api_client,
-    color_attribute_without_values,
-    permission_manage_product_types_and_attributes,
-):
-    query = UPDATE_ATTRIBUTE_MUTATION
-    attribute = color_attribute_without_values
-    name = "Wings name"
-    attribute_value_name = "Yellow Color"
-    node_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    variables = {
-        "name": name,
-        "id": node_id,
-        "addValues": [{"name": attribute_value_name}],
-        "removeValues": [],
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    get_graphql_content(response)
-    attribute.refresh_from_db()
-    assert attribute.values.count() == 1
-    assert attribute.values.filter(name=attribute_value_name).exists()
-
-
-UPDATE_ATTRIBUTE_SLUG_MUTATION = """
-    mutation updateAttribute(
-    $id: ID!, $slug: String) {
-    attributeUpdate(
-            id: $id,
-            input: {
-                slug: $slug}) {
-        errors {
-            field
-            message
-        }
-        attributeErrors {
-            field
-            message
-            code
-        }
-        attribute {
-            name
-            slug
-        }
-    }
-}
-"""
-
-
-@pytest.mark.parametrize(
-    "input_slug, expected_slug, error_message",
-    [
-        ("test-slug", "test-slug", None),
-        ("", "", "Slug value cannot be blank."),
-        (None, "", "Slug value cannot be blank."),
-    ],
-)
-def test_update_attribute_slug(
-    staff_api_client,
-    color_attribute,
-    permission_manage_product_types_and_attributes,
-    input_slug,
-    expected_slug,
-    error_message,
-):
-    query = UPDATE_ATTRIBUTE_SLUG_MUTATION
-
-    attribute = color_attribute
-    name = attribute.name
-    old_slug = attribute.slug
-
-    assert input_slug != old_slug
-
-    node_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    variables = {"slug": input_slug, "id": node_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    attribute.refresh_from_db()
-    data = content["data"]["attributeUpdate"]
-    errors = data["attributeErrors"]
-    if not error_message:
-        assert data["attribute"]["name"] == name == attribute.name
-        assert data["attribute"]["slug"] == input_slug == attribute.slug
-    else:
-        assert errors
-        assert data["attribute"] is None
-        assert errors[0]["field"] == "slug"
-        assert errors[0]["code"] == ProductErrorCode.REQUIRED.name
-
-
-def test_update_attribute_slug_exists(
-    staff_api_client, color_attribute, permission_manage_product_types_and_attributes,
-):
-    query = UPDATE_ATTRIBUTE_SLUG_MUTATION
-
-    second_attribute = Attribute.objects.get(pk=color_attribute.pk)
-    second_attribute.pk = None
-    second_attribute.slug = "second-attribute"
-    second_attribute.save()
-
-    attribute = color_attribute
-    old_slug = attribute.slug
-    new_slug = second_attribute.slug
-
-    assert new_slug != old_slug
-
-    node_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    variables = {"slug": new_slug, "id": node_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    attribute.refresh_from_db()
-    data = content["data"]["attributeUpdate"]
-    errors = data["attributeErrors"]
-
-    assert errors
-    assert data["attribute"] is None
-    assert errors[0]["field"] == "slug"
-    assert errors[0]["code"] == ProductErrorCode.UNIQUE.name
-
-
-@pytest.mark.parametrize(
-    "input_slug, expected_slug, input_name, error_message, error_field",
-    [
-        ("test-slug", "test-slug", "New name", None, None),
-        ("", "", "New name", "Slug value cannot be blank.", "slug"),
-        (None, "", "New name", "Slug value cannot be blank.", "slug"),
-        ("test-slug", "", None, "This field cannot be blank.", "name"),
-        ("test-slug", "", "", "This field cannot be blank.", "name"),
-        (None, None, None, "Slug value cannot be blank.", "slug"),
-    ],
-)
-def test_update_attribute_slug_and_name(
-    staff_api_client,
-    color_attribute,
-    permission_manage_product_types_and_attributes,
-    input_slug,
-    expected_slug,
-    input_name,
-    error_message,
-    error_field,
-):
-    query = """
-        mutation updateAttribute(
-        $id: ID!, $slug: String, $name: String) {
-        attributeUpdate(
-                id: $id,
-                input: {
-                    slug: $slug, name: $name}) {
-            errors {
-                field
-                message
-            }
-            attributeErrors {
-                field
-                message
-                code
-            }
-            attribute {
-                name
-                slug
-            }
-        }
-    }
-    """
-
-    attribute = color_attribute
-    old_name = attribute.name
-    old_slug = attribute.slug
-
-    assert input_slug != old_slug
-    assert input_name != old_name
-
-    node_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    variables = {"slug": input_slug, "name": input_name, "id": node_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    attribute.refresh_from_db()
-    data = content["data"]["attributeUpdate"]
-    errors = data["attributeErrors"]
-    if not error_message:
-        assert data["attribute"]["name"] == input_name == attribute.name
-        assert data["attribute"]["slug"] == input_slug == attribute.slug
-    else:
-        assert errors
-        assert errors[0]["field"] == error_field
-        assert errors[0]["code"] == ProductErrorCode.REQUIRED.name
-
-
-@pytest.mark.parametrize(
-    "name_1, name_2, error_msg, error_code",
-    (
-        (
-            "Red color",
-            "Red color",
-            "Provided values are not unique.",
-            ProductErrorCode.UNIQUE,
-        ),
-        (
-            "Red color",
-            "red color",
-            "Provided values are not unique.",
-            ProductErrorCode.UNIQUE,
-        ),
-    ),
-)
-def test_update_attribute_and_add_attribute_values_errors(
-    staff_api_client,
-    name_1,
-    name_2,
-    error_msg,
-    error_code,
-    color_attribute,
-    permission_manage_product_types_and_attributes,
-):
-    query = UPDATE_ATTRIBUTE_MUTATION
-    attribute = color_attribute
-    node_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    variables = {
-        "name": "Example name",
-        "id": node_id,
-        "removeValues": [],
-        "addValues": [{"name": name_1}, {"name": name_2}],
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    errors = content["data"]["attributeUpdate"]["errors"]
-    assert errors
-    assert errors[0]["field"] == "addValues"
-    assert errors[0]["message"] == error_msg
-
-    product_errors = content["data"]["attributeUpdate"]["attributeErrors"]
-    assert product_errors[0]["code"] == error_code.name
-
-
-def test_update_attribute_and_remove_others_attribute_value(
-    staff_api_client,
-    color_attribute,
-    size_attribute,
-    permission_manage_product_types_and_attributes,
-):
-    query = UPDATE_ATTRIBUTE_MUTATION
-    attribute = color_attribute
-    node_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    size_attribute = size_attribute.values.first()
-    attr_id = graphene.Node.to_global_id("AttributeValue", size_attribute.pk)
-    variables = {
-        "name": "Example name",
-        "id": node_id,
-        "addValues": [],
-        "removeValues": [attr_id],
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    errors = content["data"]["attributeUpdate"]["errors"]
-    assert errors
-    assert errors[0]["field"] == "removeValues"
-    err_msg = "Value %s does not belong to this attribute." % str(size_attribute)
-    assert errors[0]["message"] == err_msg
-
-    product_errors = content["data"]["attributeUpdate"]["attributeErrors"]
-    assert product_errors[0]["code"] == ProductErrorCode.INVALID.name
-
-
-def test_delete_attribute(
-    staff_api_client,
-    color_attribute,
-    permission_manage_product_types_and_attributes,
-    product_type,
-):
-    attribute = color_attribute
-    query = """
-    mutation deleteAttribute($id: ID!) {
-        attributeDelete(id: $id) {
-            errors {
-                field
-                message
-            }
-            attribute {
-                id
-            }
-        }
-    }
-    """
-    node_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    variables = {"id": node_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["attributeDelete"]
-    assert data["attribute"]["id"] == variables["id"]
-    with pytest.raises(attribute._meta.model.DoesNotExist):
-        attribute.refresh_from_db()
-
-
-CREATE_ATTRIBUTE_VALUE_QUERY = """
-    mutation createAttributeValue(
-        $attributeId: ID!, $name: String!) {
-    attributeValueCreate(
-        attribute: $attributeId, input: {name: $name}) {
-        attributeErrors {
-            field
-            message
-            code
-        }
-        attribute {
-            values {
-                name
-            }
-        }
-        attributeValue {
-            name
-            type
-            slug
-        }
-    }
-}
-"""
-
-
-def test_create_attribute_value(
-    staff_api_client, color_attribute, permission_manage_products
-):
-    attribute = color_attribute
-    query = CREATE_ATTRIBUTE_VALUE_QUERY
-    attribute_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    name = "test name"
-    variables = {"name": name, "attributeId": attribute_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_products]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["attributeValueCreate"]
-    assert not data["attributeErrors"]
-
-    attr_data = data["attributeValue"]
-    assert attr_data["name"] == name
-    assert attr_data["slug"] == slugify(name)
-    assert attr_data["type"] == "STRING"
-    assert name in [value["name"] for value in data["attribute"]["values"]]
-
-
-def test_create_attribute_value_not_unique_name(
-    staff_api_client, color_attribute, permission_manage_products
-):
-    attribute = color_attribute
-    query = CREATE_ATTRIBUTE_VALUE_QUERY
-    attribute_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    value_name = attribute.values.first().name
-    variables = {"name": value_name, "attributeId": attribute_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_products]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["attributeValueCreate"]
-    assert data["attributeErrors"]
-    assert data["attributeErrors"][0]["code"] == AttributeErrorCode.ALREADY_EXISTS.name
-    assert data["attributeErrors"][0]["field"] == "name"
-
-
-def test_create_attribute_value_capitalized_name(
-    staff_api_client, color_attribute, permission_manage_products
-):
-    attribute = color_attribute
-    query = CREATE_ATTRIBUTE_VALUE_QUERY
-    attribute_id = graphene.Node.to_global_id("Attribute", attribute.id)
-    value_name = attribute.values.first().name
-    variables = {"name": value_name.upper(), "attributeId": attribute_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_products]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["attributeValueCreate"]
-    assert data["attributeErrors"]
-    assert data["attributeErrors"][0]["code"] == AttributeErrorCode.ALREADY_EXISTS.name
-    assert data["attributeErrors"][0]["field"] == "name"
-
-
-UPDATE_ATTRIBUTE_VALUE_QUERY = """
-mutation updateChoice(
-        $id: ID!, $name: String!) {
-    attributeValueUpdate(
-    id: $id, input: {name: $name}) {
-        attributeErrors {
-            field
-            message
-        }
-        attributeValue {
-            name
-            slug
-        }
-        attribute {
-            values {
-                name
-            }
-        }
-    }
-}
-"""
-
-
-def test_update_attribute_value(
-    staff_api_client,
-    pink_attribute_value,
-    permission_manage_product_types_and_attributes,
-):
-    query = UPDATE_ATTRIBUTE_VALUE_QUERY
-    value = pink_attribute_value
-    node_id = graphene.Node.to_global_id("AttributeValue", value.id)
-    name = "Crimson name"
-    variables = {"name": name, "id": node_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["attributeValueUpdate"]
-    value.refresh_from_db()
-    assert data["attributeValue"]["name"] == name == value.name
-    assert data["attributeValue"]["slug"] == slugify(name)
-    assert name in [value["name"] for value in data["attribute"]["values"]]
-
-
-def test_update_attribute_value_name_not_unique(
-    staff_api_client,
-    pink_attribute_value,
-    permission_manage_product_types_and_attributes,
-):
-    query = UPDATE_ATTRIBUTE_VALUE_QUERY
-    value = pink_attribute_value.attribute.values.create(
-        name="Example Name", slug="example-name", value="#RED"
-    )
-    node_id = graphene.Node.to_global_id("AttributeValue", value.id)
-    variables = {"name": pink_attribute_value.name, "id": node_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["attributeValueUpdate"]
-    assert data["attributeErrors"]
-    assert data["attributeErrors"][0]["message"]
-    assert data["attributeErrors"][0]["field"] == "name"
-
-
-def test_delete_attribute_value(
-    staff_api_client,
-    color_attribute,
-    pink_attribute_value,
-    permission_manage_product_types_and_attributes,
-):
-    value = color_attribute.values.get(name="Red")
-    query = """
-    mutation updateChoice($id: ID!) {
-        attributeValueDelete(id: $id) {
-            attributeValue {
-                name
-                slug
-            }
-        }
-    }
-    """
-    node_id = graphene.Node.to_global_id("AttributeValue", value.id)
-    variables = {"id": node_id}
-    staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_product_types_and_attributes]
-    )
-    with pytest.raises(value._meta.model.DoesNotExist):
-        value.refresh_from_db()
 
 
 @pytest.mark.parametrize(
